removed Ballone notification. Added logic back in in.

+++ Deploy-Application.ps1
@@ -141,15 +141,22 @@ Try {
 		}
 
 		## <Perform Installation tasks here>
-		if(!(get-eventlog -LogName System -InstanceId 2147484722 -after (Get-date).addDays(-1))){
-			$OS = Get-wmiobject Win32_OperatingSystem
-			$Uptime = (Get-Date) - $OS.ConvertToDateTime($OS.LastBootUpTime)
-			[int]$DaysUp = $Uptime.TotalDays
-			Show-DialogBox -text "Your computer has been up for $DaysUp days. It needs to be restarted." -Icon Information -Timeout 900
-			   $Tomorrow = (Get-Date).AddDays(1).Date.AddHours($Time)
-			shutdown -r -t ([decimal]::round(($Tomorrow - (Get-Date)).TotalSeconds))
-			Show-InstallationRestartPrompt -CountDownSeconds ([decimal]::round(($Tomorrow - (Get-Date)).TotalSeconds))
+
+		$OS = Get-wmiobject Win32_OperatingSystem
+		$Uptime = (Get-Date) - $OS.ConvertToDateTime($OS.LastBootUpTime)
+		[int]$DaysUp = $Uptime.TotalDays
+		$Days = 7
+
+		if($DaysUp -gt $Days){
+			Write-Warning "System has been up for $DaysUp."
+    		if(!(get-eventlog -LogName System -InstanceId 2147484722 -after (Get-date).addDays(-1))){
+				Show-DialogBox -text "Your computer has been up for $DaysUp days. It needs to be restarted." -Icon Information -Timeout 900
+				$Tomorrow = (Get-Date).AddDays(1).Date.AddHours($Time)
+				shutdown -r -t ([decimal]::round(($Tomorrow - (Get-Date)).TotalSeconds))
+				Show-InstallationRestartPrompt -CountDownSeconds ([decimal]::round(($Tomorrow - (Get-Date)).TotalSeconds))
+			}
 		}
+
 		
 
 		##*===============================================
